+++ Source/SwiftLintFramework/Models/SwiftVersion.swift
@@ -9,9 +9,9 @@
 import Foundation
 import SourceKittenFramework
 
-enum SwiftVersion {
-    case three
-    case four
+enum SwiftVersion: String {
+    case three = "3"
+    case four = "4"
 
     static let current: SwiftVersion = {
         // Allow forcing the Swift version, useful in cases where SourceKit isn't available